requirements.txt
langchain==0.1.12
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
huggingface-hub==0.21.4
requests==2.31.0
pydantic==2.6.3
python-multipart==0.0.9 